Add model test
