+++ src/test/resources/xml/figure1a_updated.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<istardt  xmlns="https://example.org/istar-dt-x"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="https://example.org/istar-dt-x ../xsd/istar-rl-schema_v4.xsd">
+
+    <header title="title"
+            author="first last"
+            source="path to xml file"
+            lastUpdated="String date here">
+        notes here
+    </header>
+
+    <options continuous="true" infeasible-action-penalty="1.0"></options>
+
+</istardt>
